+++ ui/src/pages/transaction-detail.test.tsx
@@ -1,6 +1,6 @@
 import { screen } from "@testing-library/react"
 import { expect, test, vi, beforeEach } from "vitest"
-import { MemoryRouter, Route, Routes } from "react-router-dom"
+import { MemoryRouter, Route, Routes } from "react-router"
 import { renderWithI18n } from "@/test/i18n-test-utils"
 import { TransactionDetailPage } from "./transaction-detail"